requirements.txt
opencv-python-headless
pytesseract
discord.py
